+++ packages/components/pages/f-checkout/test/visual/f-checkout-guest.visual.mobile.spec.js
@@ -12,39 +12,39 @@ describe('f-checkout - Collection - Guest - Mobile Visual Tests', () => {
         locale: 'en-GB'
     };
 
-    beforeEach(() => {
+    beforeEach(async () => {
         // Arrange
         checkout = new Checkout();
 
         // Act
-        checkout.load({
+        await checkout.load({
             ...checkoutInfo
         });
     });
 
-    it('should display the component base state.', () => {
+    it('should display the component base state.', async () => {
         // Assert
-        browser.percyScreenshot('f-checkout - Collection - Guest - Base State', 'mobile');
+        await browser.percyScreenshot('f-checkout - Collection - Guest - Base State', 'mobile');
     });
 
-    it('should display the mandatory error messages', () => {
+    it('should display the mandatory error messages', async () => {
         // Act
-        checkout.goToPayment();
+        await checkout.goToPayment();
 
         // Assert
-        browser.percyScreenshot('f-checkout - Collection - Guest - Mandatory Errors State', 'mobile');
+        await browser.percyScreenshot('f-checkout - Collection - Guest - Mandatory Errors State', 'mobile');
     });
 
-    it('should display the illegal mobile number error message', () => {
+    it('should display the illegal mobile number error message', async () => {
         // Act
-        checkout.setFieldValue('mobileNumber', illegalMobileNumber);
-        checkout.goToPayment();
+        await checkout.setFieldValue('mobileNumber', illegalMobileNumber);
+        await checkout.goToPayment();
 
         // Assert
-        browser.percyScreenshot('f-checkout - Collection - Guest - Illegal Mobile Number Error State', 'mobile');
+        await browser.percyScreenshot('f-checkout - Collection - Guest - Illegal Mobile Number Error State', 'mobile');
     });
 
-    it('should display the "Duplicate Order Warning" modal', () => {
+    it('should display the "Duplicate Order Warning" modal', async () => {
         // Arrange
         checkout = new Checkout();
 
@@ -58,25 +58,26 @@ describe('f-checkout - Collection - Guest - Mobile Visual Tests', () => {
         checkout.inputFieldValues = customerInfo;
 
         // Act
-        checkout.load({
+        await checkout.load({
             ...checkoutInfo,
             placeOrderError: 'duplicate'
         });
-        checkout.setFieldValues();
-        checkout.goToPayment();
+        await checkout.setFieldValues();
+        await browser.pause(200);
+        await checkout.goToPayment();
 
         // Assert
-        browser.percyScreenshot('f-checkout - Collection - Guest - "Duplicate Order Warning" Modal', 'mobile');
+        await browser.percyScreenshot('f-checkout - Collection - Guest - "Duplicate Order Warning" Modal', 'mobile');
     });
 });
 
 describe('f-checkout - Collection - Guest - isAsapAvailable: false Mobile Visual Tests', () => {
-    beforeEach(() => {
+    beforeEach(async () => {
         // Arrange
         checkout = new Checkout();
 
         // Act
-        checkout.load({
+        await checkout.load({
             serviceType: 'collection',
             isLoggedIn: false,
             isAsapAvailable: false,
@@ -84,9 +85,9 @@ describe('f-checkout - Collection - Guest - isAsapAvailable: false Mobile Visual
         });
     });
 
-    it('should display the pre-order warning.', () => {
+    it('should display the pre-order warning.', async () => {
         // Assert
-        browser.percyScreenshot('f-checkout - Collection - Guest - Pre-Order Warning', 'mobile');
+        await browser.percyScreenshot('f-checkout - Collection - Guest - Pre-Order Warning', 'mobile');
     });
 });
 
@@ -99,39 +100,39 @@ describe('f-checkout - Delivery - Guest - Mobile Visual Tests', () => {
         locale: 'en-GB'
     };
 
-    beforeEach(() => {
+    beforeEach(async () => {
         // Arrange
         checkout = new Checkout();
 
         // Act
-        checkout.load({
+        await checkout.load({
             ...checkoutInfo
         });
     });
 
-    it('should display the delivery f-checkout component guest base state.', () => {
+    it('should display the delivery f-checkout component guest base state.', async () => {
         // Assert
-        browser.percyScreenshot('f-checkout - Delivery - Guest - Base State', 'mobile');
+        await browser.percyScreenshot('f-checkout - Delivery - Guest - Base State', 'mobile');
     });
 
-    it('should display the delivery f-checkout guest mandatory error messages', () => {
+    it('should display the delivery f-checkout guest mandatory error messages', async () => {
         // Act
-        checkout.goToPayment();
+        await checkout.goToPayment();
 
         // Assert
-        browser.percyScreenshot('f-checkout - Delivery - Guest - Mandatory Errors State', 'mobile');
+        await browser.percyScreenshot('f-checkout - Delivery - Guest - Mandatory Errors State', 'mobile');
     });
 
-    it('should display the illegal mobile number error message', () => {
+    it('should display the illegal mobile number error message', async () => {
         // Act
-        checkout.setFieldValue('mobileNumber', illegalMobileNumber);
-        checkout.goToPayment();
+        await checkout.setFieldValue('mobileNumber', illegalMobileNumber);
+        await checkout.goToPayment();
 
         // Assert
-        browser.percyScreenshot('f-checkout - Delivery - Guest - Illegal Mobile Number Error State', 'mobile');
+        await browser.percyScreenshot('f-checkout - Delivery - Guest - Illegal Mobile Number Error State', 'mobile');
     });
 
-    it('should display the "Duplicate Order Warning" modal', () => {
+    it('should display the "Duplicate Order Warning" modal', async () => {
         // Arrange
         checkout = new Checkout();
 
@@ -148,25 +149,26 @@ describe('f-checkout - Delivery - Guest - Mobile Visual Tests', () => {
         checkout.inputFieldValues = customerInfo;
 
         // Act
-        checkout.load({
+        await checkout.load({
             ...checkoutInfo,
             placeOrderError: 'duplicate'
         });
-        checkout.setFieldValues();
-        checkout.goToPayment();
+        await checkout.setFieldValues();
+        await browser.pause(200);
+        await checkout.goToPayment();
 
         // Assert
-        browser.percyScreenshot('f-checkout - Delivery - Guest - "Duplicate Order Warning" Modal', 'mobile');
+        await browser.percyScreenshot('f-checkout - Delivery - Guest - "Duplicate Order Warning" Modal', 'mobile');
     });
 });
 
 describe('f-checkout - Delivery - Guest - isAsapAvailable: false Mobile Visual Tests', () => {
-    beforeEach(() => {
+    beforeEach(async () => {
         // Arrange
         checkout = new Checkout();
 
         // Act
-        checkout.load({
+        await checkout.load({
             serviceType: 'delivery',
             isLoggedIn: false,
             isAsapAvailable: false,
@@ -174,9 +176,9 @@ describe('f-checkout - Delivery - Guest - isAsapAvailable: false Mobile Visual T
         });
     });
 
-    it('should display the pre-order warning.', () => {
+    it('should display the pre-order warning.', async () => {
         // Assert
-        browser.percyScreenshot('f-checkout - Delivery - Guest - Pre-Order Warning', 'mobile');
+        await browser.percyScreenshot('f-checkout - Delivery - Guest - Pre-Order Warning', 'mobile');
     });
 });
 
@@ -188,40 +190,40 @@ describe('f-checkout - Dine In - Guest - Mobile Visual Tests', () => {
         locale: 'en-GB'
     };
 
-    beforeEach(() => {
+    beforeEach(async () => {
         // Arrange
         checkout = new Checkout();
 
         // Act
-        checkout.load({
+        await checkout.load({
             ...checkoutInfo
         });
     });
 
-    it('should display the component base state.', () => {
+    it('should display the component base state.', async () => {
         // Assert
-        browser.percyScreenshot('f-checkout - Dine in - Guest - Base State', 'mobile');
+        await browser.percyScreenshot('f-checkout - Dine in - Guest - Base State', 'mobile');
     });
 
-    it('should display the mandatory error messages', () => {
+    it('should display the mandatory error messages', async () => {
         // Act
-        checkout.clearBlurField('tableIdentifier');
-        checkout.goToPayment();
+        await checkout.clearBlurField('tableIdentifier');
+        await checkout.goToPayment();
 
         // Assert
-        browser.percyScreenshot('f-checkout - Dine In - Guest - Manadatory Errors', 'mobile');
+        await browser.percyScreenshot('f-checkout - Dine In - Guest - Manadatory Errors', 'mobile');
     });
 
-    it('should display the illegal mobile number error message', () => {
+    it('should display the illegal mobile number error message', async () => {
         // Act
-        checkout.setFieldValue('mobileNumber', illegalMobileNumber);
-        checkout.goToPayment();
+        await checkout.setFieldValue('mobileNumber', illegalMobileNumber);
+        await checkout.goToPayment();
 
         // Assert
-        browser.percyScreenshot('f-checkout - Dine In - Guest - Illegal Mobile Number Error State', 'mobile');
+        await browser.percyScreenshot('f-checkout - Dine In - Guest - Illegal Mobile Number Error State', 'mobile');
     });
 
-    it('should display the "Duplicate Order Warning" modal', () => {
+    it('should display the "Duplicate Order Warning" modal', async () => {
         // Arrange
         checkout = new Checkout();
 
@@ -236,14 +238,15 @@ describe('f-checkout - Dine In - Guest - Mobile Visual Tests', () => {
         checkout.inputFieldValues = customerInfo;
 
         // Act
-        checkout.load({
+        await checkout.load({
             ...checkoutInfo,
             placeOrderError: 'duplicate'
         });
-        checkout.setFieldValues();
-        checkout.goToPayment();
+        await checkout.setFieldValues();
+        await browser.pause(200);
+        await checkout.goToPayment();
 
         // Assert
-        browser.percyScreenshot('f-checkout - Dine in - Guest - "Duplicate Order Warning" Modal', 'mobile');
+        await browser.percyScreenshot('f-checkout - Dine in - Guest - "Duplicate Order Warning" Modal', 'mobile');
     });
 });