+++ packages/components/pages/f-checkout/test/visual/f-checkout-invalid.visual.mobile.spec.js
@@ -14,8 +14,7 @@ describe('f-checkout - Invalid - Mobile Visual Tests', () => {
         checkout = new Checkout();
     });
 
-
-    it('should display the error page component if the url is invalid', () => {
+    it('should display the error page component if the url is invalid', async () => {
         // Act
         checkout.load({
             ...checkoutInfo,
@@ -23,10 +22,10 @@ describe('f-checkout - Invalid - Mobile Visual Tests', () => {
         });
 
         // Assert
-        browser.percyScreenshot('f-checkout - Invalid - Base State', 'mobile');
+        await browser.percyScreenshot('f-checkout - Invalid - Base State', 'mobile');
     });
 
-    it('should display the "Get Checkout 403" error page', () => {
+    it('should display the "Get Checkout 403" error page', async () => {
         // Act
         checkout.load({
             ...checkoutInfo,
@@ -35,10 +34,10 @@ describe('f-checkout - Invalid - Mobile Visual Tests', () => {
         });
 
         // Assert
-        browser.percyScreenshot('f-checkout - Delivery - Authenticated - "Get Checkout 403" Error Page', 'mobile');
+        await browser.percyScreenshot('f-checkout - Delivery - Authenticated - "Get Checkout 403" Error Page', 'mobile');
     });
 
-    it('should display an error dialog if the basket has invalid products', () => {
+    it('should display an error dialog if the basket has invalid products', async () => {
         // Act
         checkout.load({
             ...checkoutInfo,
@@ -47,10 +46,10 @@ describe('f-checkout - Invalid - Mobile Visual Tests', () => {
         });
 
         // Assert
-        browser.percyScreenshot('f-checkout - Basket error: Invalid products - Base State', 'mobile');
+        await browser.percyScreenshot('f-checkout - Basket error: Invalid products - Base State', 'mobile');
     });
 
-    it('should display an error dialog if the basket has offline products', () => {
+    it('should display an error dialog if the basket has offline products', async () => {
         // Act
         checkout.load({
             ...checkoutInfo,
@@ -59,10 +58,10 @@ describe('f-checkout - Invalid - Mobile Visual Tests', () => {
         });
 
         // Assert
-        browser.percyScreenshot('f-checkout - Basket error: Offline products - Base State', 'mobile');
+        await browser.percyScreenshot('f-checkout - Basket error: Offline products - Base State', 'mobile');
     });
 
-    it('should display an error dialog if age verification is required', () => {
+    it('should display an error dialog if age verification is required', async () => {
         checkout.load({
             ...checkoutInfo,
             serviceType: 'delivery',
@@ -70,6 +69,6 @@ describe('f-checkout - Invalid - Mobile Visual Tests', () => {
         });
 
         // Assert
-        browser.percyScreenshot('f-checkout - Basket error: Age restriction required - Base State', 'mobile');
+        await browser.percyScreenshot('f-checkout - Basket error: Age restriction required - Base State', 'mobile');
     });
 });