+++ packages/components/pages/f-contact-preferences/test/component/f-contact-preferences.component.spec.js
@@ -5,62 +5,63 @@ const ContactPreferences = require('../../test-utils/component-objects/f-contact
 let contactPreferences;
 
 describe('f-contact-preferences component tests', () => {
-    beforeEach(() => {
+    beforeEach(async () => {
         // Arrange
         contactPreferences = new ContactPreferences();
     });
 
     forEach([
         ['en-GB']
-    ]).it('should display the  %s f-contact-preferences component', locale => {
+    ]).it('should display the  %s f-contact-preferences component', async locale => {
         // Act
-        contactPreferences.load({ locale });
+        await contactPreferences.load({ locale });
 
         // Assert
-        expect(contactPreferences.isComponentDisplayed()).toBe(true);
+        await expect(await contactPreferences.isComponentDisplayed()).toBe(true);
     });
 
     forEach([
         ['en-GB']
-    ]).it('should display the %s Submit success alert if Submit succeed', locale => {
+    ]).it('should display the %s Submit success alert if Submit succeed', async locale => {
         // Act
-        contactPreferences.load({ locale });
-        contactPreferences.clickNewsEmailCheckbox(); // dirty the form to allow submit
-        contactPreferences.clickSubmitButton();
-        contactPreferences.waitForComponent();
+        await contactPreferences.load({ locale });
+        await contactPreferences.clickNewsEmailCheckbox(); // dirty the form to allow submit
+        await contactPreferences.clickSubmitButton();
+        await contactPreferences.waitForComponent();
 
         // Assert
-        expect(contactPreferences.isErrorAlertDisplayed()).toBe(false);
-        expect(contactPreferences.isSuccessAlertDisplayed()).toBe(true);
+        await expect(await contactPreferences.isErrorAlertDisplayed()).toBe(false);
+        await expect(await contactPreferences.isSuccessAlertDisplayed()).toBe(true);
     });
 
     forEach([
         ['en-GB']
-    ]).it('should display the %s Submit error alert if Submit fails', locale => {
+    ]).it('should display the %s Submit error alert if Submit fails', async locale => {
         // Act
-        contactPreferences.load({
+        await contactPreferences.load({
             locale,
             apiState: 'post-preferences-fails'
         });
-        contactPreferences.clickNewsEmailCheckbox(); // dirty the form to allow submit
-        contactPreferences.clickSubmitButton();
-        contactPreferences.waitForComponent();
+        // browser.debug();
+        await contactPreferences.clickNewsEmailCheckbox(); // dirty the form to allow submit
+        await contactPreferences.clickSubmitButton();
+        await contactPreferences.waitForComponent();
 
         // Assert
-        expect(contactPreferences.isSuccessAlertDisplayed()).toBe(false);
-        expect(contactPreferences.isErrorAlertDisplayed()).toBe(true);
+        await expect(await contactPreferences.isSuccessAlertDisplayed()).toBe(false);
+        await expect(await contactPreferences.isErrorAlertDisplayed()).toBe(true);
     });
 
     forEach([
         ['en-GB']
-    ]).it('should display the %s Error page if Load fails', locale => {
+    ]).it('should display the %s Error page if Load fails', async locale => {
         // Act
-        contactPreferences.load({
+        await contactPreferences.load({
             locale,
             apiState: 'get-preferences-fails'
         });
 
         // Assert
-        expect(contactPreferences.isErrorPageDisplayed()).toBe(true);
+        await expect(await contactPreferences.isErrorPageDisplayed()).toBe(true);
     });
 });