+++ packages/components/pages/f-registration/src/components/Registration.vue
@@ -422,7 +422,7 @@ export default {
         },
         password: {
             required,
-            minLength: minLength(4)
+            minLength: minLength(10)
         }
     },