+++ packages/components/pages/f-registration/CHANGELOG.md
@@ -3,6 +3,20 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+Latest (add to next release)
+------------------------------
+*June 23, 2021*
+
+### Changed
+- Refactor WebDriverIO tests to use async in order to support Node 16 using `codemod` utility.
+
+
+v2.4.2
+------------------------------
+*June 29, 2022*
+
+### Added
+- Increase password minimum length from 4 to 10 characters.
 
 v2.4.1
 ------------------------------
@@ -34,6 +48,10 @@ v2.2.0
 
 ### Changed
 - Bumped wdio version and fixed breaking changes.
+
+*February 4, 2022*
+
+### Changed
 - Upgraded to ESLint v8